+++ src/test/kotlin/io/mustelidae/riverotter/domain/calendar/api/HolidayCalendarControllerTest.kt
@@ -5,7 +5,7 @@ import io.kotlintest.shouldBe
 import io.kotlintest.shouldNotBe
 import io.mustelidae.riverotter.common.AvailableCountry
 import io.mustelidae.riverotter.domain.calendar.holiday.repository.HolidayCalendarRepository
-import io.mustelidae.riverotter.domain.config.IntegrationSupport
+import io.mustelidae.riverotter.config.IntegrationSupport
 import org.junit.jupiter.api.MethodOrderer
 import org.junit.jupiter.api.Order
 import org.junit.jupiter.api.Test
