+++ src/test/kotlin/io/mustelidae/riverotter/domain/topic/api/TopicHolidayCalendarControllerTest.kt
@@ -5,7 +5,7 @@ import io.kotlintest.shouldBe
 import io.mustelidae.riverotter.domain.calendar.api.HolidayCalendarController
 import io.mustelidae.riverotter.domain.calendar.api.HolidayCalendarResources
 import io.mustelidae.riverotter.domain.calendar.holiday.Holiday
-import io.mustelidae.riverotter.domain.config.IntegrationSupport
+import io.mustelidae.riverotter.config.IntegrationSupport
 import io.mustelidae.riverotter.domain.topic.Workday
 import org.junit.jupiter.api.BeforeAll
 import org.junit.jupiter.api.Test
